+++ src/api/handlers/collection/write/create.rs
@@ -37,6 +37,8 @@ impl ContentService {
         let mut data: DocumentFields = req
             .data
             .map(|s| data_map_to_json_map(&s))
+            .transpose()
+            .map_err(Status::invalid_argument)?
             .unwrap_or_default()
             .into();