+++ src/hooks/lua_api/mod.rs
@@ -52,5 +52,5 @@ pub(crate) mod utils;
 mod vm_label;
 
 pub use register::{register_api, register_api_pool_init};
-pub(crate) use serializers::{json_to_lua, lua_to_json, set_max_nesting_depth};
+pub(crate) use serializers::{json_to_lua, lua_to_json, max_nesting_depth, set_max_nesting_depth};
 pub use vm_label::VmLabel;